GCW-3036 addressed PR comments
