+++ app/constants/item-actions.js
@@ -1,0 +1,18 @@
+export const ITEM_ACTIONS = [
+  {
+    name: "Process",
+    icon: "random"
+  },
+  {
+    name: "Recycle",
+    icon: "recycle"
+  },
+  {
+    name: "Trash",
+    icon: "dumpster"
+  },
+  {
+    name: "Loss",
+    icon: "folder-minus"
+  }
+];
