+++ app/services/package-service.js
@@ -1,4 +1,5 @@
 import ApiBaseService from "./api-base-service";
+import { toID } from "stock/utils/helpers";
 
 export default ApiBaseService.extend({
   store: Ember.inject.service(),
@@ -90,5 +91,32 @@ export default ApiBaseService.extend({
       });
     });
     return packagesTypes;
+  },
+
+  /**
+   * Performs action on a package from the specified location
+   *
+   * @param {Package} pkg the package to move
+   * @param {object} opts the move properties
+   * @param {Location|string} opts.from the source location or its id
+   * @param {quantity} opts.quantity the quantity to move
+   * @param {comment} opts.comment the comment of package action
+   * @returns {Promise<Model>}
+   */
+  async peformActionOnPackage(pkg, opts = {}) {
+    const { from, actionName, quantity, comment } = opts;
+
+    const payload = await this.PUT(
+      `/packages/${pkg.get("id")}/actions/${opts.actionName}`,
+      {
+        quantity,
+        from: toID(from),
+        description: comment
+      }
+    );
+
+    this.get("store").pushPayload(payload);
+
+    return this.get("store").peekRecord("item", pkg.get("id"));
   }
 });
